backend/requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-CORS==4.0.0
python-dotenv==1.0.0
Werkzeug==2.3.0
beautifulsoup4==4.12.0
requests==2.31.0
pytest==7.4.0
pytest-cov==4.1.0
gunicorn==21.2.0
